pyproject.toml
[project]
name = "fu7ur3pr00f"
version = "0.2.0"
description = "Career intelligence system - gather professional data, generate CVs, analyze career alignment"
readme = "README.md"
requires-python = ">=3.13"
license = "GPL-2.0-only"
authors = [
    {name = "Juan Manuel Daza"},
]
keywords = ["career", "cv", "resume", "AI", "LLM", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    # Core
    "typer>=0.15",
    # click<8.3 required: cli.py patches make_metavar() for Typer 0.24 compat.
    # Remove upper bound once Typer supports Click 8.3+ natively.
    "click>=8.1,<8.3",
    "pydantic-settings>=2",
    "pyyaml>=6",
    "rich>=14",
    # AI/LLM
    "langchain>=0.3",
    "langgraph>=0.3",
    "langchain-openai>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-google-genai>=2",
    "langchain-ollama>=0.3",
    # Data gathering
    "httpx>=0.28",
    "beautifulsoup4>=4",
    "defusedxml>=0.7",
    "nh3>=0.2",  # HTML sanitization
    # MCP (Model Context Protocol)
    "mcp>=1.26",
    # Generation
    "weasyprint>=62",
    "markdown>=3.10.1",
    "python-jobspy>=1.1.82",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "textual>=0.80.0",
    "terminaltexteffects>=0.14",
    "chromadb>=1.4.1",
    # Testing
    "pytest-asyncio>=0.23",
    "pytest-textual-snapshot",
]

[project.urls]
Homepage = "https://github.com/juanmanueldaza/fu7ur3pr00f"
Repository = "https://github.com/juanmanueldaza/fu7ur3pr00f"
"Bug Tracker" = "https://github.com/juanmanueldaza/fu7ur3pr00f/issues"

[project.scripts]
fu7ur3pr00f = "fu7ur3pr00f.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fu7ur3pr00f"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["--strict-markers", "--strict-config", "--verbose", "--tb=short"]
markers = [
    "integration: marks tests as integration tests (DAG: multiple components wired together)",
    "unit: marks tests as unit tests (mocked external services, single component)",
    "slow: marks tests as slow (benchmarks, large iterations)",
]
